+++ packages/core/src/metrics/internal.ts
@@ -12,7 +12,7 @@ import { timestampInSeconds } from '../utils/time';
 import { _getTraceInfoFromScope } from '../utils/trace-info';
 import { createMetricEnvelope } from './envelope';
 
-const MAX_METRIC_BUFFER_SIZE = 100;
+const MAX_METRIC_BUFFER_SIZE = 1000;
 
 /**
  * Converts a metric attribute to a serialized metric attribute.
@@ -172,7 +172,7 @@ export function _INTERNAL_captureMetric(beforeMetric: Metric, options?: Internal
 
   if (replayId && replay?.getRecordingMode() === 'buffer') {
     // We send this so we can identify cases where the replayId is attached but the replay itself might not have been sent to Sentry
-    setMetricAttribute(processedMetricAttributes, 'sentry._internal.replay_is_buffering', replayId);
+    setMetricAttribute(processedMetricAttributes, 'sentry._internal.replay_is_buffering', true);
   }
 
   const metric: Metric = {